requirements.txt
flask
flask-cors
python-dotenv
supabase
gunicorn
requests
